Logged "all languages" in correct language
The "all language" was hardcoded in the database in the language of
the user who submitted it, so it was displayed in the wrong language
to users with a different user language. This commit adds an empty
string as parameter, and changes that into the correct user language
in the log.

Patch set 2: Added comments and a wrapped the commit message to make
             it Gerrit-friendly.

Change-Id: I0238df8d6280c7bc4e1e66067aa8e3a68c8dd4a9