Add a new job to send out emails using UserMailer directly
No longer using EmaillingJob to send out email notifications as it
requires non-serializable parameters that don't work well with a
kafka based queue.

Added TranslationNotificationEmailJob and creating an instance of
that in the TranslationNotifyUser class.

Renamed the TranslationNotificationJob to
TranslationNotificationTalkPageJob since it is used to leave messages
on user's talk page.

Bug: T144780
Change-Id: Ie56fae91ca630a72b27aa488556a6be90b1e11de