Remove TranslationNotificationsTalkPageJob
It's purpose was to figure out the target page name for cross wiki
delivery. However, it doesn't work because TranslationNotifications
is not enabled on except a few wikis so they do not know how to process
this job.

Simplify the code using the assumption that user names are global and
construct MassMessageJobs directly.

Bug: T144780
Change-Id: I64e482e465dc6b8fe1f9958e1af60d8616c7e20b