Special:NotifyTranslators: Use entity selector for selecting message groups
Loading of message groups using the select field displays a lot of
results that are hard to seatch. This fix uses the EntitySelector
to speed up the selection by allowing the user to enter a search term
that is then debounced.

The selection of discouraged translatable pages is disallowed to keep
functionality similar to the non-JavaScript experience.

Add a default empty option to the translatable page selector, and
ensure that one of the translatable page options is selected before
sending notification.

Bug: T253984
Depends-on: I16da60e7e8f9d5bdbb6a33059486af23b6c8e782
Change-Id: Iaceb8e46ae7418612fd959c79a5dd78076c14060